Move scale zoom methods, move more methods to helper
