+++ src/geo/projection/globe_utils.ts
@@ -1,9 +1,9 @@
 import {vec3} from 'gl-matrix';
-import {clamp, lerp, mod, remapSaturate, wrap} from '../../util/util';
+import {clamp, lerp, mod, remapSaturate, scaleZoom, wrap} from '../../util/util';
 import {LngLat} from '../lng_lat';
-import {MAX_VALID_LATITUDE, scaleZoom} from '../transform_helper';
-import type Point from '@mapbox/point-geometry';
+import {MAX_VALID_LATITUDE} from '../transform_helper';
 import {EXTENT} from '../../data/extent';
+import type Point from '@mapbox/point-geometry';
 
 export function getGlobeCircumferencePixels(transform: {worldSize: number; center: {lat: number}}): number {
     const radius = getGlobeRadiusPixels(transform.worldSize, transform.center.lat);
