+++ src/geo/projection/mercator_utils.ts
@@ -1,11 +1,12 @@
 import {mat4} from 'gl-matrix';
 import {EXTENT} from '../../data/extent';
-import {type OverscaledTileID} from '../../source/tile_id';
-import {clamp, createIdentityMat4f32, degreesToRadians} from '../../util/util';
-import {MAX_VALID_LATITUDE, type UnwrappedTileIDType, zoomScale} from '../transform_helper';
-import {type LngLat} from '../lng_lat';
+import {clamp, createIdentityMat4f32, degreesToRadians, zoomScale} from '../../util/util';
+import {MAX_VALID_LATITUDE, type UnwrappedTileIDType} from '../transform_helper';
 import {MercatorCoordinate, mercatorXfromLng, mercatorYfromLat, mercatorZfromAltitude} from '../mercator_coordinate';
 import Point from '@mapbox/point-geometry';
+
+import type {OverscaledTileID} from '../../source/tile_id';
+import type {LngLat} from '../lng_lat';
 import type {ProjectionData} from './projection_data';
 
 /*
